Create Base_Genetic
We use the yf.download() function from yfinance to fetch the Apple stock price data for the last 5 years. The data is then stored in the data DataFrame. We reset the index to ensure that the dates are included as a column.

The genetic algorithm trading strategy runs on the fetched stock price data. Note that the yfinance library relies on internet connectivity to fetch the data from Yahoo Finance. Ensure that system has internet access while running this code.